move handleEvalOut to its own package (#635)
This function is moved into its own package so arr.ai binaries can import this function
